Extract ScheduleService from schedule controller
