+++ convergence_games/apps/frontend/admin/services/__init__.py
@@ -1,0 +1,6 @@
+from ._schedule_service import ScheduleService, provide_schedule_service
+
+__all__ = [
+    "ScheduleService",
+    "provide_schedule_service",
+]
